refactor(automations): extract automation lists
